+++ src/js/store/index.js
@@ -1,6 +1,6 @@
 import Vue from 'vue';
 import Vuex from 'vuex';
-// import fetchPosts from '../Services/posts';
+import fetchPosts from '../Services/posts';
 
 Vue.use(Vuex);
 
@@ -53,26 +53,7 @@ export default new Vuex.Store({
       context.commit('addToActivityLog', payload.message);
     },
     fetchPosts(context) {
-      return new Promise((resolve, reject) => {
-        context.commit('setError', null);
-        context.commit('isLoadingPosts');
-
-        const url = 'https://jsonplaceholder.typicode.com/posts';
-
-        fetch(url)
-          .then(response => response.json())
-          .then(posts => {
-            context.commit('isNotLoadingPosts');
-            context.commit('addPostsToStore', posts);
-            context.commit('addToActivityLog', `Data retrieved from ${url}`);
-            resolve(posts);
-          })
-          .catch(error => {
-            context.commit('setError', error);
-            context.commit('addToActivityLog', `ERROR fetching data from ${url}. See log for details.`);
-            reject(error);
-          });
-      });
+      return fetchPosts(context);
     }
   },
   getters: {